(minor) FS: Removed quick take/wear code from guild manage screen as new guild store has this functionality now.
(minor) FS: Enhanced link from world screen to the realm on UFSG so it goes directly to the map.

(minor) FS: Fixed bug on the quest book pages that fetches the next step of quest data so it shows up correctly now.
(minor) FS: Removed some unused code on the guild manage page that blocked the page jump down to the guild store.